Updated 04_calc_Sheet1_fluxes
#Problem:

There is an error occurred that area files are not found.

#Solution:

The code is updated for creating a stat folder inside the data folder. Without this update, the output folder will not be generated inside the data folder. It'll raise an error. Now, the Area_km.tif and Basin_Area_km.tif files are created inside the data folder.